generalization of Course Basics
CourseNumber and course name is properties shared between
Curriculum plan and course info (from schedule)
so we can put that in a base class.
