+++ src/main/java/auto_scheduler/CurriculumCourse.java
@@ -1,25 +1,7 @@
 package org.txstate.auto_scheduler;
 
-public class CurriculumCourse {
-    private String courseNumber;
-    private String courseName;
+public class CurriculumCourse extends CourseBase {
     private String preRequisite;
-
-    public String getCourseNumber() {
-        return this.courseNumber;
-    }
-
-    public void setCourseNumber(String courseNumber) {
-        this.courseNumber = courseNumber;
-    }
-
-    public String getCourseName() {
-        return this.courseName;
-    }
-
-    public void setCourseName(String courseName) {
-        this.courseName = courseName;
-    }
 
     public String getPreRequisite() {
         return this.preRequisite;
